backend/requirements.txt
# VimaERP 2.0 - Backend Dependencies
# Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# ORM + Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Validation & Settings
pydantic[email]==2.10.4
pydantic-settings==2.7.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Background Tasks
celery[redis]==5.4.0
redis==5.2.1

# HTTP Client
httpx==0.28.1

# Utils
python-dotenv==1.0.1
python-multipart==0.0.20
python-ulid==3.1.0

# Production Server
gunicorn==23.0.0
pandas==2.2.3
openpyxl==3.1.5
xlrd==2.0.1

